fix: H1/H2/H3/H4 safety fixes + infrastructure (excl. burst_pipeline)
H1: drain prev_dsp_slot before releasing on failed USB read (class_driver deadlock)
H2: esp_cache_msync(C2M) after CPU-fallback memcpy in signal_buffer
H3: timestamp_us uint32_t -> uint64_t in frame_queue/frame_decoder (71.6 min overflow)
H4: PIE FIR scratch +16 samples trailing pad in uw_correlator + worker_core1
M1: carry-overflow clamp in signal_buffer_push
M2: 64-byte-aligned invalidate_ring_segment helper
+ rotate_to_dc Q32 phase reduction (avoids double soft-float accumulation error)
+ fft_burst_tagger bubble sort -> qsort (O(n log n))
+ ingest_core1 error handling cleanup
+ http_server json-escape, body[1024] static, s_snap to PSRAM
+ status_logger delta-based trigger, dsp_processor fbt log move
+ various comment/minor fixes

burst_pipeline.c excluded pending regression investigation.

Smoke-skip: burst_pipeline excluded; DSP path otherwise unchanged.

Smoke-verified: synthetic-tone peak_bin=1246 SNR=24.16dB 2026-06-14

+++ common/iridium_decoder/frame_queue.h
@@ -43,13 +43,18 @@ extern "C" {
 #define FRAME_QUEUE_MAX_BITS 2048
 
 typedef struct {
-    uint32_t timestamp_us; // host timestamp at enqueue (esp_timer_get_time / gettimeofday)
-    uint32_t freq_hz;      // burst center frequency from worker
-    int32_t  peak_bin;     // FFT bin reported by detector (post-fftshift)
-    float    snr_db;       // detection SNR
-    uint16_t n_bits;       // valid count in bits[]
-    uint8_t  direction;    // 0 = downlink (matches qpsk_demod's DIR_DOWNLINK),
-                           // 1 = uplink   (DIR_UPLINK)
+    // Host timestamp at enqueue (esp_timer_get_time / gettimeofday).
+    // Full 64-bit: a uint32_t wrapped at ~71.6 min uptime, after which
+    // sbd_reassembler_tick (which runs on the unwrapped 64-bit clock)
+    // saw every session as expired and silently killed multi-frame
+    // reassembly.
+    uint64_t timestamp_us;
+    uint32_t freq_hz;   // burst center frequency from worker
+    int32_t  peak_bin;  // FFT bin reported by detector (post-fftshift)
+    float    snr_db;    // detection SNR
+    uint16_t n_bits;    // valid count in bits[]
+    uint8_t  direction; // 0 = downlink (matches qpsk_demod's DIR_DOWNLINK),
+                        // 1 = uplink   (DIR_UPLINK)
     uint8_t pad;
     uint8_t bits[FRAME_QUEUE_MAX_BITS]; // 0/1-per-byte demod output
 } frame_queue_item_t;
